+++ src/main/java/site/mylittlestore/dto/item/ItemCreationDto.java
@@ -13,15 +13,15 @@ public class ItemCreationDto {
 
     private String name;
 
-    private int price;
+    private Long price;
 
-    private int stock;
+    private Long stock;
 
     private String image;
 
     @Builder
     @QueryProjection
-    public ItemCreationDto(Long id, Long storeId, String name, int price, int stock, String image) {
+    public ItemCreationDto(Long id, Long storeId, String name, Long price, Long stock, String image) {
         this.id = id;
         this.storeId = storeId;
         this.name = name;
